+++ src/lib/k8s.ts
@@ -2,6 +2,7 @@ import { Err, Ok, Result } from "./result";
 import { SrchdError } from "./error";
 
 import * as k8s from "@kubernetes/client-node";
+import { Writable } from "stream";
 
 export const kc = new k8s.KubeConfig();
 kc.loadFromDefault();
@@ -257,5 +258,116 @@ export async function timeout(
         ),
       );
     }, timeoutMs);
+  });
+}
+
+export async function podExec(
+  cmd: string[],
+  workspaceId: string,
+  computerId?: string,
+  timeoutMs?: number,
+): Promise<
+  Result<{ stdout: string; stderr: string; exitCode: number }, SrchdError>
+> {
+  const k8sExec = new k8s.Exec(kc);
+  let stdout = "";
+  let stderr = "";
+  let exitCode = 0;
+  let failReason: "commandRunFailed" | "executionFailed" | undefined;
+  const execPromise = new Promise<void>((resolve, reject) => {
+    const stdoutStream = new Writable({
+      write(chunk, _enc, callback) {
+        stdout += chunk.toString();
+        callback();
+      },
+    });
+
+    const stderrStream = new Writable({
+      write(chunk, _encoding, callback) {
+        stderr += chunk.toString();
+        callback();
+      },
+    });
+
+    k8sExec
+      .exec(
+        workspaceId,
+        podName(workspaceId, computerId),
+        computerId ? "computer" : "srchd",
+        cmd,
+        stdoutStream,
+        stderrStream,
+        null,
+        false,
+        (status: k8s.V1Status) => {
+          stdoutStream.end();
+          stderrStream.end();
+
+          /* Error Status example:
+          {
+            "metadata": {},
+            "status": "Failure",
+            "message": "command terminated with non-zero exit code: command terminated with exit code 127",
+            "reason": "NonZeroExitCode",
+            "details": {
+              "causes": [
+                {
+                  "reason": "ExitCode",
+                  "message": "127"
+                }
+              ]
+            }
+          }
+          */
+
+          if (status.status === "Success") {
+            exitCode = 0;
+          } else {
+            reject(new Error(status.message));
+            const tryToNumber = Number(status.details?.causes?.[0]?.message);
+            exitCode = isNaN(tryToNumber) ? 1 : tryToNumber;
+            failReason = "commandRunFailed";
+          }
+          resolve();
+        },
+      )
+      .catch((err: any) => {
+        failReason = "executionFailed";
+        reject(new Error(err));
+      });
+  });
+
+  try {
+    if (!timeoutMs) {
+      await execPromise;
+    } else {
+      await Promise.race([execPromise, timeout(timeoutMs)]);
+    }
+  } catch (err: any) {
+    if (failReason === "commandRunFailed") {
+      return new Ok({
+        exitCode,
+        stdout,
+        stderr,
+      });
+    }
+
+    if (err instanceof Err) {
+      return err;
+    }
+
+    return new Err(
+      new SrchdError(
+        "pod_run_error",
+        `Failed to execute ${cmd.join(" ")}`,
+        new Error(err),
+      ),
+    );
+  }
+
+  return new Ok({
+    exitCode,
+    stdout,
+    stderr,
   });
 }
